+++ tests/test_cases/objective_c/cpp_header/input/widget.h
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace app {
+
+class Widget {
+public:
+    void run();
+    int value() const;
+};
+
+}